example/requirements.txt
asgiref==3.8.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bootstrap4==0.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
debugpy==1.6.0
decorator==5.1.1
Django==4.0.6
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==22.1
django-pandas==0.6.6
django-request==1.6.3
django-simple-history==3.5.0
django-tables2==2.4.1
docutils==0.19
entrypoints==0.4
executing==0.8.3
factory-boy==3.3.0
Faker==25.0.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
keyring==23.6.0
matplotlib-inline==0.1.3
names==0.3.0
nest-asyncio==1.5.5
numpy==1.26.4
packaging==21.3
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
pip==24.0
pkginfo==1.8.3
plotly==5.9.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==23.2.0
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tablib==3.6.1
tenacity==8.0.1
tornado==6.2
tqdm==4.66.2
traitlets==5.3.0
twine==4.0.1
tzdata==2024.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
wincertstore==0.2
zipp==3.8.0
